requirements.txt
Django==5.0.6
djangorestframework==3.15.2
gunicorn==23.0.0
dj-database-url==2.2.0
whitenoise==6.7.0
psycopg2==2.9.9
asgiref==3.8.1
tzdata==2024.1